회원 장바구니에 물건 안담기는 문제 수정

+++ backend/src/main/java/com/example/backend/product/dto/CartItemDto.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class CartItemDto {
-    private Long productId;
+    private Integer productId;
     private String optionName;
     private Integer quantity;
 }
